pip_requirements.txt
coverage
coloredlogs
click>=6.7
flake8
flask
flask-babel
flask-login
flask-mail
flask-openid
flask-sqlalchemy
flask-whooshalchemy
flask-wtf
ipython>=5.4.1
jupyter>=1.0.0
matplotlib
numpy
pylint
plotly
pep8
pytest>=3.2
pytest-cov>=2.5.1
pytest-runner>=2.11.1
pytest-mock
progressbar33
setuptools
simplejson>=3.11.1
Sphinx>=1.6.3
sphinx_rtd_theme>=0.2.4
tox
pyaml>=16.12.2
tensorflow==1.0
tensorflow-tensorboard
keras
